Create SearchImage.py
First draft of image scanning with all extraneous stuff removed.

New stockpile name will have to be handled.  Left the submit to bot code in there and commented out so it can be replaced with whatever will happen to perhaps modify the database directly.  Function SearchImage should be passed an image path.  Threading appeared to break things as commenting it out and just passing a singular image worked fine.  Since this script will be called each time an image is imported, is it inherently threaded?  Last two lines of code are there for testing purposes.  Currently prints the result array.  Successfully tested internally with both full screenshot and cropped screenshot.